fix: handle properties with dot in their name
In `walk`, the current `path` is represented as a string, which causes
all kinds on problems when the linter tries to work with it (eg. extract
a path segment). This commit changes that, and makes `path` an array of
path segments, so it's easier to manipulate, eg. it's straighforward to
handle properties with exotic names (contaning a `.`).

In some cases (when a non-empty `path` was given to `walk`), the
resulting path was inconsistent, so `pathToString` was added that will
transform the `path` array to a consistent (ie. jq-compatible) json path
expression.